% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_references.R
\name{CheckReferences}
\alias{CheckReferences}
\title{Loads a context definition and checks its syntax.}
\usage{
CheckReferences(con)
}
\arguments{
\item{con}{a connection containing the meta data}
}
\description{
Loads a context definition and checks its syntax.
}
\examples{
filePath <- system.file("extdata", "context1.yaml", package="datapR")
errors <- CheckReferences(filePath)
print(errors)

}

