% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Combine}
\alias{Combine}
\title{Combines two xts objects with similar data,
giving preference to objects first in the list.}
\usage{
Combine(listofxts)
}
\arguments{
\item{listofxts}{xts objects to combine}
}
\description{
Combines two xts objects with similar data,
giving preference to objects first in the list.
}
\examples{
library(Quandl)
ubs1 <- Quandl("GOOG/VTX_UBSN", type = "xts")
ubs2 <- ubs1
ubs3 <- ubs1

#assume ubs1 and ubs2 and ubs3 come from different
#data sources, each of which may have missing data.
#We simulate by messing with the data:

ubs1 <- ubs1[-c(3,5), ]
ubs2 <- ubs2[-3, ]
ubs3[1, 1] <- 58

#Source 1 is considered best, but sometimes
#it may have missing data which is available in 2
#or 3.
ubs <- Combine(list(ubs1, ubs2, ubs3))

}

