% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ConstantReturn}
\alias{ConstantReturn}
\title{Create a daily timeseries with a constant yearly return.}
\usage{
ConstantReturn(startDate = "1990-01-01", endDate = Sys.Date(), r,
  startValue = 1, dayCountConvention = c("weekdays", "calendardays",
  "holidays"), holiday = NULL)
}
\arguments{
\item{startDate}{the first date of the xts to be returned}

\item{endDate}{the last date of the xts to be returned}

\item{r}{the yearly return}

\item{startValue}{the first value of the xts}

\item{dayCountConvention}{the convention to be used}

\item{holiday}{for convention \code{holidays}, a \code{timeDate} holiday calendar}
}
\value{
an \code{xts} object
}
\description{
Create a daily timeseries with a constant yearly return.
}

