% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datap.R
\docType{package}
\name{datap}
\alias{datap}
\alias{datap-package}
\title{Data aquisition and pre-processing using datap contexts}
\description{
datap is a DSL to acquire and pre-process data from various sources.
}
\examples{
library(datap)
filePath <- system.file("extdata", "context1.yaml", package="datap")
context <- Load(filePath)

context$`Closing Prices`$Indices$SPX
plot(context$`Closing Prices`$Indices$SPX)
spx <- context$`Closing Prices`$Indices$SPX$tap()
spxFullRange <- context$`Closing Prices`$Indices$SPX$tap(dteRange = "/")


}

