% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapR.R
\docType{package}
\name{datapR}
\alias{datapR}
\alias{datapR-package}
\title{Data aquisition and pre-processing using datap contexts}
\description{
datap is a DSL to acquire and pre-process data from various sources.
datapR is a datap interpreter for R.
}
\examples{
library(datapR)
filePath <- system.file("extdata", "context1.yaml", package="datapR")
context <- Load(filePath)
context
context$`Closing Prices`$Indices$SPX
plot(context$`Closing Prices`$Indices$SPX)
context$`Closing Prices`$Indices$SPX$tap()
context$`Closing Prices`$Indices$SPX$tap(dteRange = "/")


}

