% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FixedWeightIndex}
\alias{FixedWeightIndex}
\title{Generate an index in which each component has a fixed weight.}
\usage{
FixedWeightIndex(weights, children)
}
\description{
Generate an index in which each component has a fixed weight.
}

