% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{Get}
\alias{Get}
\title{Get a timeseries}
\usage{
Get(code, db, ...)
}
\arguments{
\item{code}{the code of the timeseries}

\item{db}{the db object}

\item{...}{and additional parameters that need to be passed to the function}
}
\description{
Get a timeseries
}

