% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{GetData}
\alias{GetData}
\title{Get a timeseries}
\usage{
GetData(id, context, ...)
}
\arguments{
\item{id}{the id of the timeseries}

\item{context}{the context object}

\item{...}{and additional parameters that need to be passed to the function}
}
\description{
Get a timeseries
}

