% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{GetDynamicVariables}
\alias{GetDynamicVariables}
\title{Get variables that are dynamic, such
as @pipe, @pipefun or @context. Assumes non-dynamic
variables have already been parsed.}
\usage{
GetDynamicVariables(node)
}
\description{
Get variables that are dynamic, such
as @pipe, @pipefun or @context. Assumes non-dynamic
variables have already been parsed.
}

