% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{GetUpstreamParameters}
\alias{GetUpstreamParameters}
\title{Finds the tap parameters that will be used on
this node or on any of its children}
\usage{
GetUpstreamParameters(node)
}
\description{
Finds the tap parameters that will be used on
this node or on any of its children
}

