% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{Load}
\alias{Load}
\title{Loads a meta definition.}
\usage{
Load(con)
}
\arguments{
\item{con}{a connection containing the meta data}
}
\description{
Loads a meta definition.
}
\examples{
filePath <- system.file("extdata", "sample_metadata.yaml", package="finPrice")
context <- Load(filePath)

}

