% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings_errors.R
\name{NaRatio}
\alias{NaRatio}
\title{Check the ratio of NA values compared to non-NA values.}
\usage{
NaRatio(timeseries, variable, maxRatio)
}
\arguments{
\item{timeseries}{the timeseries object (can be mostly any class)}

\item{variable}{the name of the column}

\item{maxRatio}{the maximum ratio}
}
\description{
Check the ratio of NA values compared to non-NA values.
}

