% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Ones}
\alias{Ones}
\title{Create a timeseries of 1}
\usage{
Ones(startDate = "1990-01-01", endDate = Sys.Date())
}
\arguments{
\item{startDate}{the first date of the xts to be returned}

\item{endDate}{the last date of the xts to be returned}
}
\description{
Create a timeseries of 1
}

