% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ParseVariables}
\alias{ParseVariables}
\title{Finds static variables in arguments and replaces them
with their values.}
\usage{
ParseVariables(node)
}
\description{
Finds static variables in arguments and replaces them
with their values.
}

