% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.specification}
\alias{plot.specification}
\title{Plot a timeseries or a module}
\usage{
\method{plot}{specification}(x, ..., direction = c("climb", "descend"),
  pruneFun = NULL, engine = "dot")
}
\description{
Plot a timeseries or a module
}
\examples{
filePath <- system.file("extdata", "sample_metadata.yaml", package="finPrice")
context <- Load(filePath)
plot(context$SPX)
plot(context$`NA Handling`)

}

