% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{plot.timeseries}
\alias{plot.timeseries}
\title{Plot a timeseries (or a context)}
\usage{
\method{plot}{timeseries}(x, ..., direction = c("climb", "descend"),
  pruneFun = NULL, engine = "dot")
}
\description{
Plot a timeseries (or a context)
}
\examples{
filePath <- system.file("extdata", "sample_metadata.yaml", package="finPrice")
context <- Load(filePath)
plot(context$SPX)

}

