% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Regularize}
\alias{Regularize}
\title{Regularize the index of a daily xts by adding NAs to
non-existent days}
\usage{
Regularize(xts)
}
\arguments{
\item{xts}{the xts object to regularize}
}
\description{
Regularize the index of a daily xts by adding NAs to
non-existent days
}
\examples{
library(Quandl)
library(magrittr)
Quandl("GOOG/VTX_UBSN", type = "xts", force_irregular = TRUE) \%>\%
Regularize \%>\% na.locf -> ubs

}

