% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SetNames}
\alias{SetNames}
\title{Set the names to an object}
\usage{
SetNames(x, names)
}
\arguments{
\item{x}{the object on which to set the names}

\item{names}{the names}
}
\description{
Set the names to an object
}

