% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Tap}
\alias{Tap}
\title{Get data from a tap}
\usage{
Tap(context, tapPath, ...)
}
\arguments{
\item{context}{the context object}

\item{tapPath}{the name of the tap}

\item{...}{any parameters to be passed on to the tap}
}
\description{
Get data from a tap
}

