% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Cache}
\alias{Cache}
\title{Cache / memoise a function for some time}
\usage{
Cache(joint, timeoutSec)
}
\arguments{
\item{joint}{the joint whose upstream joint function should be cached}

\item{timeoutSec}{the number of seconds ot cache}
}
\description{
Cache / memoise a function for some time
}

