% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ForgetCache}
\alias{ForgetCache}
\title{Forget memoised/cached function}
\usage{
ForgetCache(joint, inflow)
}
\arguments{
\item{joint}{the joint whose downstream joint function should be forgotten}

\item{inflow}{a value passed as a result}
}
\description{
Forget memoised/cached function
}

