% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Ones}
\alias{Ones}
\title{Create a timeseries of 1}
\usage{
Ones(startDate = "1990-01-01", endDate = Sys.Date(), colname = "Close")
}
\arguments{
\item{startDate}{the start date}

\item{endDate}{the end date}

\item{colname}{the name the series should get}
}
\description{
Create a timeseries of 1
}

