% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tap}
\alias{plot.tap}
\title{Plot a timeseries or a module}
\usage{
\method{plot}{tap}(x, ..., pruneFun = NULL)
}
\arguments{
\item{x}{the tap to plot}

\item{...}{additional parameters to pass to the plot.Node method}

\item{pruneFun}{a function to prune the plot. See \code{\link[data.tree]{Prune}}}
}
\description{
Plot a timeseries or a module
}
\examples{
filePath <- system.file("extdata", "context1.yaml", package="datap")
context <- Load(filePath)
plot(data.tree::FindNode(context, 'SPX'))

}

