% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{Tap}
\alias{Tap}
\title{Get data from a tap}
\usage{
Tap(context, tapName, ...)
}
\arguments{
\item{context}{the context object}

\item{tapName}{the name of the tap}

\item{...}{any parameters to be passed on to the tap}
}
\description{
Get data from a tap
}

